+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Adminer dev
+- Fix search anywhere (bug #1004, regression from 5.1.1)
 
 ## Adminer 5.2.0 (released 2025-04-08)
 - Autocomplete SQL commands