+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Adminer dev
 - Fix search anywhere (bug #1004, regression from 5.1.1)
 - Fix import without primary key (bug #1017, regression from 5.1.1)
+- PostgreSQL PDO: Fix bytea without primary key (bug #1021)
 
 ## Adminer 5.2.0 (released 2025-04-08)
 - Autocomplete SQL commands